challenge/pyproject.toml
[project]
name = "challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "langchain>=1.2.8",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "openai>=2.16.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.14",
    "yt-dlp>=2026.1.31",
]
